refactor(rendering): state the persistent sprite slot rules once
Both backends had grown their own copy of the same two loops: the
acquisition walk that decides whether one slot store can serve a source
(no own material, one blend mode, every base texture inside one table) and
the ENTER write that copies prepacked rows into the slots that were just
handed out. Neither loop has a backend-specific line in it — the only input
that differs is the batch table's width.

Move both to `persistentSpriteSlots`, against a structural store interface
rather than a base class: the two stores share no representation at all, one
addressing its slots through data textures and the other through storage
buffers, and a common ancestor would claim otherwise.
