+++ src/rendering/webgpu/WebGpuSpriteRenderer.ts
@@ -12,7 +12,7 @@ import {
 } from '#rendering/material/RetainedMaterialState';
 import type { SpriteMaterial } from '#rendering/material/SpriteMaterial';
 import type { RenderRootSource } from '#rendering/plan/RenderRootSource';
-import { SOURCE_QUAD_FLOATS } from '#rendering/sourceQuadRecord';
+import { fillPersistentSpriteSlotTable, writePersistentSpriteSlots } from '#rendering/sprite/persistentSpriteSlots';
 import type { Sprite } from '#rendering/sprite/Sprite';
 import {
   buildSpriteTextureSlotWgsl,
@@ -25,7 +25,6 @@ import {
 import { DataTexture } from '#rendering/texture/DataTexture';
 import { RenderTexture } from '#rendering/texture/RenderTexture';
 import { Texture } from '#rendering/texture/Texture';
-import { TRANSFORM_FLOATS_PER_ROW, TRANSFORM_TINT_BYTES_PER_ROW } from '#rendering/TransformBuffer';
 import { BlendModes } from '#rendering/types';
 import type { View } from '#rendering/View';
 
@@ -621,20 +620,10 @@ export class WebGpuSpriteRenderer extends AbstractWebGpuRenderer<Sprite> impleme
    * Decide whether this renderer can serve every item in `source` from one slot
    * store, and allocate it if so.
    *
-   * Three refusals, each of them a batching rule rather than an ordering one —
-   * ordering was already settled by the plan layer:
-   *
-   * - A sprite carrying its own {@link SpriteMaterial} takes the custom path,
-   *   where a material switch is a hard pass/pipeline boundary. One ordered
-   *   stream cannot express that, so any own-material item disqualifies the
-   *   source.
-   * - A blend-mode change is likewise a pipeline boundary, so the whole source
-   *   has to agree on one.
-   * - The base textures must all fit ONE slot table. That is what makes a slot
-   *   texture index stable for the item's whole life — with one table there is
-   *   no per-batch re-slotting for a membership change to invalidate.
-   *
-   * Runs once per built source, never per frame.
+   * The rules are the shared sprite ones (see
+   * {@link fillPersistentSpriteSlotTable}); what is backend-specific is only the
+   * batch table's width and the store the answer is written into. Runs once per
+   * built source, never per frame.
    * @internal
    */
   public _acquirePersistentSlotStore(source: RenderRootSource, backend: WebGpuBackend): WebGpuPersistentSlotStore | null {
@@ -644,79 +633,24 @@ export class WebGpuSpriteRenderer extends AbstractWebGpuRenderer<Sprite> impleme
       return null;
     }
 
-    this._ensurePersistentResources(device);
-
     const store = new WebGpuPersistentSlotStore();
-    const textureIndexOfHandle = new Uint8Array(source.itemCount);
-    let blendMode: BlendModes | null = null;
 
-    for (const scope of source.scopes) {
-      const drawables = scope.items.drawables;
-      const count = scope.items.count;
-      const handleBase = scope.handleBase;
-
-      for (let i = 0; i < count; i++) {
-        const sprite = drawables[i] as Sprite;
-        const texture = sprite.texture;
-
-        if (texture === null || sprite.material !== null) {
-          store.destroy();
-
-          return null;
-        }
-
-        const mode = sprite.blendMode;
-
-        if (blendMode === null) {
-          blendMode = mode;
-        } else if (blendMode !== mode) {
-          store.destroy();
-
-          return null;
-        }
-
-        let index = store.textures.indexOf(texture);
-
-        if (index === -1) {
-          if (store.textures.length >= this._maxBatchTextures) {
-            store.destroy();
-
-            return null;
-          }
-
-          index = store.textures.push(texture) - 1;
-        }
-
-        // Recorded per item here, on the walk that builds the table anyway, so
-        // an ENTER never has to ask a drawable which texture it uses.
-        textureIndexOfHandle[handleBase + i] = index;
-      }
-    }
-
-    if (blendMode === null) {
+    if (!fillPersistentSpriteSlotTable(source, store, this._maxBatchTextures)) {
       store.destroy();
 
       return null;
     }
 
-    store.blendMode = blendMode;
-    store.textureIndexOfHandle = textureIndexOfHandle;
+    this._ensurePersistentResources(device);
     store.connectDevice(device, backend.accountant);
 
     return store;
   }
 
   /**
-   * Fill the persistent rows of the items that just took a slot, entirely from
-   * the source's prepacked tables.
-   *
-   * `entered` is a flat `(scopeOrdinal, localIndex, slot)` triple list holding
-   * arrivals only — a staying item's rows are already what this would write,
-   * which is the whole saving. No drawable is touched: an item entering the view
-   * has not been read for hundreds of frames, and resolving its transform,
-   * bounds and texture frame out of cold objects was the measured cost of a
-   * camera step. Everything here is a fixed-size copy between typed arrays, plus
-   * one lookup for the store-table texture index.
+   * Fill the persistent rows of the items that just took a slot and push the
+   * blocks they landed on. See {@link writePersistentSpriteSlots} for why this
+   * never touches a drawable.
    * @internal
    */
   public _writePersistentSlotRows(store: WebGpuPersistentSlotStore, source: RenderRootSource, entered: Int32Array, count: number): void {
@@ -726,44 +660,13 @@ export class WebGpuSpriteRenderer extends AbstractWebGpuRenderer<Sprite> impleme
       return;
     }
 
-    const scopes = source.scopes;
-    const textureIndexOfHandle = store.textureIndexOfHandle;
-    let highest = -1;
-
-    for (let i = 0; i < count; i++) {
-      const slot = entered[i * 3 + 2]!;
-
-      if (slot > highest) {
-        highest = slot;
-      }
-    }
-
     // Both the growth (which replaces the GPU buffers) and the block uploads
-    // below land on the queue timeline ahead of the deferred submit, so a draw
-    // of THIS store already recorded into the open pass would silently pick them
-    // up. End that pass first; a pass holding anyone else's draws is unaffected,
-    // because nothing but this store's own draws reads these buffers.
+    // land on the queue timeline ahead of the deferred submit, so a draw of THIS
+    // store already recorded into the open pass would silently pick them up. End
+    // that pass first; a pass holding anyone else's draws is unaffected, because
+    // nothing but this store's own draws reads these buffers.
     this._endPassOnPersistentHazard(backend, store);
-    store.ensureCapacity(highest + 1);
-
-    for (let i = 0; i < count; i++) {
-      const base = i * 3;
-      const scope = scopes[entered[base]!]!;
-      const localIndex = entered[base + 1]!;
-      const items = scope.items;
-
-      store.writeSlotFrom(
-        entered[base + 2]!,
-        items.rows,
-        localIndex * TRANSFORM_FLOATS_PER_ROW,
-        items.tints,
-        localIndex * TRANSFORM_TINT_BYTES_PER_ROW,
-        items.quads,
-        localIndex * SOURCE_QUAD_FLOATS,
-        textureIndexOfHandle[scope.handleBase + localIndex]!,
-      );
-    }
-
+    writePersistentSpriteSlots(store, source, entered, count);
     store.commitDirtySlots();
   }
